Move property dict validator to separate utility file
